third_party/WebKit/LayoutTests/crypto/digest-failures.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests incorrect calls to crypto.subtle.digest()");

jsTestIsAsync = true;

// Called with too few parameters.
shouldThrow("crypto.subtle.digest({name: 'sha-1'})");

Promise.resolve(null).then(function(result) {
    // "null" is not a valid data argument.
    return crypto.subtle.digest({name: 'sha-1'}, null);
}).then(failAndFinishJSTest, function(result) {
    error = result;
    shouldBeNull("error");

    // 10 is not a valid data argument.
    return crypto.subtle.digest({name: 'sha-1'}, 10);
}).then(failAndFinishJSTest, function(result) {
    error = result;
    shouldBeNull("error");

    // null is not a valid algorithm argument.
    data = new Uint8Array([0]);
    return crypto.subtle.digest(null, data);
}).then(failAndFinishJSTest, function(result) {
    error = result;
    shouldBeNull("error");

    // "sha" is not a recognized algorithm name
    return crypto.subtle.digest({name: 'sha'}, data);
}).then(failAndFinishJSTest, function(result) {
    error = result;
    shouldBeNull("error");

    // Algorithm lacks a name.
    return crypto.subtle.digest({}, data);
}).then(failAndFinishJSTest, function(result) {
    error = result;
    shouldBeNull("error");
  
}).then(finishJSTest, failAndFinishJSTest);

</script>

</body>
</html>
